test(_debug): split tests by source module into test_debug_args.py
Mirror source layout: tests/_debug/test_debug_args.py covers get_debug_args + field validators; tests/_debug/test_underflow_overflow.py covers check_compile_conflict and UnderflowOverflowMonitor. Locks in the trace_batch_nums non-negative validator. ruff check + format --check + mypy + 11/11 pytest green.
