+++ tests/_debug/test_underflow_overflow.py
@@ -16,9 +16,7 @@ from lightning_utilities.core.imports import RequirementCache
 from lightly_train._debug.debug_args import (
     DebugArgs,
     DebugUnderflowOverflowArgs,
-    get_debug_args,
 )
-from lightly_train.errors import ConfigValidationError
 
 if not RequirementCache("transformers"):
     pytest.skip("Transformers not installed", allow_module_level=True)
@@ -89,23 +87,6 @@ def make_monitor(
 
     for monitor in created:
         monitor.close()
-
-
-def test_get_debug_args__parses_user_config() -> None:
-    assert get_debug_args(None).is_underflow_overflow_enabled() is False
-
-    args = get_debug_args(
-        {"underflow_overflow": {"enabled": True, "max_frames_to_save": 5}}
-    )
-
-    assert args.is_underflow_overflow_enabled() is True
-    assert args.underflow_overflow is not None
-    assert args.underflow_overflow.max_frames_to_save == 5
-
-
-def test_get_debug_args__rejects_unknown_keys() -> None:
-    with pytest.raises(ConfigValidationError):
-        get_debug_args({"underflow_overflow": {"nonexistent_key": 1}})
 
 
 def test_check_compile_conflict__disabled_debug_allows_compile() -> None:
